organise project in seperate files
